+++ editors/zed/languages/ppphp/brackets.scm
@@ -1,3 +1,6 @@
 ("{" @open "}" @close)
 ("[" @open "]" @close)
 ("(" @open ")" @close)
+
+(type_arguments "<" @open ">" @close)
+(type_parameters "<" @open ">" @close)